Move parseEventServerErrors function to separate file and move most important functions to the to by using function declarations
